requirements/base.txt
Django>=4.2,<4.3
django-dotenv==1.4.1
wagtail>=5.1,<5.2
wagtail-font-awesome-svg>=0.0.3,<1
django-debug-toolbar>=3.2,<4
django-extensions==3.2.1
django-csp==3.7
dj-database-url==0.4.1
django-allauth<=0.56.1
